use size_t for length in API of byte_*()
There should be no code that ever comes close to this size, but it will not cost
us anything on most platforms.

+++ byte_cr.c
@@ -1,6 +1,6 @@
 #include "byte.h"
 
-void byte_copyr(char *to, unsigned int n, const char *from)
+void byte_copyr(char *to, size_t n, const char *from)
 {
   to += n;
   from += n;
